preview modal complete
