+++ pkg/web/scene/scene.go
@@ -9,7 +9,6 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/trisacrypto/envoy/pkg"
 	"github.com/trisacrypto/envoy/pkg/config"
-	"github.com/trisacrypto/envoy/pkg/store/models"
 	"github.com/trisacrypto/envoy/pkg/web/api/v1"
 	"github.com/trisacrypto/envoy/pkg/web/auth"
 )
@@ -161,7 +160,6 @@ func (s Scene) AccountPerson() Person {
 			return Person{
 				Forename:       account.FirstName,
 				Surname:        account.LastName,
-				FullLegalName:  account.FullName(),
 				CustomerNumber: account.CustomerNumber(),
 			}
 		}
@@ -192,48 +190,6 @@ func (s Scene) UserDetail() *api.User {
 func (s Scene) CounterpartyList() *api.CounterpartyList {
 	if data, ok := s[APIData]; ok {
 		if out, ok := data.(*api.CounterpartyList); ok {
-			return out
-		}
-	}
-	return nil
-}
-
-func (s Scene) TransactionsList() *TransactionList {
-	if data, ok := s[APIData]; ok {
-		if txns, ok := data.(*api.TransactionsList); ok {
-			out := &TransactionList{
-				Page:         txns.Page,
-				Transactions: make([]*Transaction, len(txns.Transactions)),
-			}
-
-			for i, txn := range txns.Transactions {
-				out.Transactions[i] = &Transaction{
-					Transaction: *txn,
-					Status:      Status(txn.Status),
-				}
-			}
-
-			return out
-		}
-	}
-	return nil
-}
-
-func (s Scene) TransactionDetail() *Transaction {
-	if data, ok := s[APIData]; ok {
-		if tx, ok := data.(*api.Transaction); ok {
-			return &Transaction{
-				Transaction: *tx,
-				Status:      Status(tx.Status),
-			}
-		}
-	}
-	return nil
-}
-
-func (s Scene) TransactionCounts() *models.TransactionCounts {
-	if data, ok := s[APIData]; ok {
-		if out, ok := data.(*models.TransactionCounts); ok {
 			return out
 		}
 	}
